Rewrite rendering.
New Renderer class used to render objects.
Adjustments to classes to fit to the new design.
New Stack class and EmptyStackException class.
New GroupRenderable class which allows several renderables to be grouped and transformed together.
Rendering now uses a pattern similar to that of a visitor pattern. A Renderer object is passed to Renderables to render with.
Projection is no longer immutable.
Removal of obsolete classes.